+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.1] - 2017-12-15
 - Convert Closure `Object` to TypeScript `object`.
 - Use glob patterns instead of RegExps to exclude files.
+- Bump Analyzer version to include https://github.com/Polymer/polymer-analyzer/pull/791 which makes Polymer properties possibly `null|undefined`.
 
 ## [0.3.0] - 2017-12-12
 - `void` is not nullable.